# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 家庭 AI 学习助手 API
  version: 0.6.0
  description: Self-hosted Household account/session and learning contract vertical slices.
servers:
  - url: http://localhost:8000
paths:
...
  /households/{household_id}/tasks:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
    get:
      operationId: listTasks
      summary: List tasks visible to the caller
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Tasks visible to the parent or bound child.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudyTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createTask
      summary: Create a math task for a child in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyTask'
        '200':
          description: Idempotent task-create replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/tasks/{task_id}/sessions:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/TaskId'
    post:
      operationId: startStudySession
      summary: Start a session for a task assigned to the bound child
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartStudySessionRequest'
      responses:
        '201':
          description: Study session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudySession'
        '200':
          description: Idempotent start-session replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudySession'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/sessions/{session_id}/attempts:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/SessionId'
    post:
      operationId: recordAttempt
      summary: Append an attempt to the bound child's active study session
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordAttemptRequest'
      responses:
        '201':
          description: Attempt appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '200':
          description: Idempotent attempt replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attempt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/sync-batches:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
    post:
      operationId: syncOfflineBatch
      summary: Apply a versioned batch of offline attempt events atomically
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatchRequest'
      responses:
        '200':
          description: Batch applied or replayed without partial writes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
...
components:
  parameters:
    HouseholdId:
      name: household_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    BearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque 256-bit session token
      description: Flutter session token; only a SHA-256 digest is stored by the API.
    SessionCookie:
      type: apiKey
      in: cookie
      name: study_session
      description: HttpOnly SameSite=Lax Web session cookie.
  schemas:
...
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required: [child_id, title, subject, scheduled_for]
      properties:
        child_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 120
        subject:
          type: string
          enum: [math]
        scheduled_for:
          type: string
          format: date
    StudyTask:
      type: object
      additionalProperties: false
      required: [id, household_id, child_id, title, subject, scheduled_for, status, version, created_at]
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        title:
          type: string
        subject:
          type: string
          enum: [math]
        scheduled_for:
          type: string
          format: date
        status:
          type: string
          enum: [assigned, in_progress]
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    StartStudySessionRequest:
      type: object
      additionalProperties: false
      required: [expected_task_version]
      properties:
        expected_task_version:
          type: integer
          minimum: 1
    StudySession:
      type: object
      additionalProperties: false
      required: [id, household_id, child_id, task_id, task_version, status, started_at]
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        task_version:
          type: integer
          minimum: 1
        status:
          type: string
          const: active
        started_at:
          type: string
          format: date-time
...
    RecordAttemptRequest:
      type: object
      additionalProperties: false
      required: [event_id, answer_summary]
      properties:
        event_id:
          type: string
          format: uuid
        answer_summary:
          type: string
          minLength: 1
          maxLength: 200
    Attempt:
      type: object
      additionalProperties: false
      required: [id, event_id, household_id, child_id, session_id, sequence, answer_summary, recorded_at]
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 1
        answer_summary:
          type: string
        recorded_at:
          type: string
          format: date-time
    SyncEvent:
      type: object
      additionalProperties: false
      required: [event_id, idempotency_key, kind, session_id, answer_summary]
      properties:
        event_id:
          type: string
          format: uuid
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 128
        kind:
          type: string
          const: record_attempt
        session_id:
          type: string
          format: uuid
        answer_summary:
          type: string
          minLength: 1
          maxLength: 200
    SyncBatchRequest:
      type: object
      additionalProperties: false
      required: [schema_version, events]
      properties:
        schema_version:
          type: integer
          const: 1
        events:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: '#/components/schemas/SyncEvent'
    SyncEventResult:
      type: object
      additionalProperties: false
      required: [event_id, status, attempt]
      properties:
        event_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [applied, replayed]
        attempt:
          $ref: '#/components/schemas/Attempt'
    SyncBatchResult:
      type: object
      additionalProperties: false
      required: [schema_version, results]
      properties:
        schema_version:
          type: integer
          const: 1
        results:
          type: array
          items:
            $ref: '#/components/schemas/SyncEventResult'
  responses:
    Unauthorized:
      description: Missing or invalid synthetic principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Principal role cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource is not visible in the caller's household.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdempotencyConflict:
      description: Idempotency key was reused with a different payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...